Moved Scheduler to a separate file. Started implementing actual scheduling, but it needs more thinking
