+++ src/Beast/Common/Types.h
@@ -1,17 +1,7 @@
 #pragma once
-#include "Beast/Common/Id.h"
-
 #include <memory>
-#include <string_view>
-#include <filesystem>
 #include <cstdint>
-#include <string>
-#include <map>
-#include <unordered_map>
 #include <chrono>
-#include <map>
-#include <string>
-#include <vector>
 
 namespace be
 {
