+++ tests/Unit/Beast/CMakeLists.txt
@@ -21,7 +21,7 @@ add_executable(
   "Input/InputTest.cpp"
   "Ecs/WorldTest.cpp"
   "Ecs/ViewTest.cpp"
-)
+ "Ecs/SchedulerTest.cpp")
 target_include_directories(BeastCoreUnits PRIVATE "${BeastEngine_SOURCE_DIR}/tests")
 
 find_package(GTest REQUIRED)
